pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "isaacsim-bench"
version = "0.1.0"
description = "Synthetic benchmark for Isaac Sim asset retrieval and scene assembly"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "numpy",
    "scipy",
    "click",
    "Pillow",
    "usd-core>=26.3",
    "open-clip-torch>=3.3.0",
    "torch>=2.11.0",
]

[project.scripts]
isaacsim-bench = "isaacsim_bench.cli.main:cli"

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ["src/isaacsim_bench"]

[tool.pytest.ini_options]
testpaths = ["tests"]
